+++ app/javascript/mastodon/components/column.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import scrollTop from '../scroll';
 
 class Column extends React.PureComponent {
 
@@ -7,11 +8,33 @@ class Column extends React.PureComponent {
     children: PropTypes.node,
   };
 
+  scrollTop () {
+    const scrollable = this.node.querySelector('.scrollable');
+
+    if (!scrollable) {
+      return;
+    }
+
+    this._interruptScrollAnimation = scrollTop(scrollable);
+  }
+
+  handleWheel = () => {
+    if (typeof this._interruptScrollAnimation !== 'function') {
+      return;
+    }
+
+    this._interruptScrollAnimation();
+  }
+
+  setRef = c => {
+    this.node = c;
+  }
+
   render () {
     const { children } = this.props;
 
     return (
-      <div role='region' className='column'>
+      <div role='region' className='column' ref={this.setRef} onWheel={this.handleWheel}>
         {children}
       </div>
     );
