+++ app/javascript/mastodon/features/community_timeline/index.js
@@ -54,6 +54,10 @@ class CommunityTimeline extends React.PureComponent {
     dispatch(moveColumn(columnId, dir));
   }
 
+  handleHeaderClick = () => {
+    this.column.scrollTop();
+  }
+
   componentDidMount () {
     const { dispatch, streamingAPIBaseURL, accessToken } = this.props;
 
@@ -98,18 +102,23 @@ class CommunityTimeline extends React.PureComponent {
     }
   }
 
+  setRef = c => {
+    this.column = c;
+  }
+
   render () {
     const { intl, hasUnread, columnId } = this.props;
     const pinned = !!columnId;
 
     return (
-      <Column>
+      <Column ref={this.setRef}>
         <ColumnHeader
           icon='users'
           active={hasUnread}
           title={intl.formatMessage(messages.title)}
           onPin={this.handlePin}
           onMove={this.handleMove}
+          onClick={this.handleHeaderClick}
           pinned={pinned}
         />
